Renamed atomix-atomic to atomix-primitives. This was necessary since atomix supports serializable consistency and therefore the word "atomic" was misleading.
