--- atomic/pom.xml
@@ -1,41 +1,0 @@
-<!--
-  ~ Copyright 2015 the original author or authors.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>io.atomix</groupId>
-    <artifactId>atomix-parent</artifactId>
-    <version>0.1.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>atomix-atomic</artifactId>
-  <name>Atomix Atomic Variables</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>io.atomix</groupId>
-      <artifactId>atomix-resource</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.atomix.catalyst</groupId>
-      <artifactId>catalyst-local</artifactId>
-      <version>${catalyst.version}</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-</project>
