+++ examples/pom.xml
@@ -29,7 +29,7 @@
     <module>standalone-server</module>
     <module>leader-election</module>
     <module>group-membership</module>
-    <module>atomic-value</module>
+    <module>distributed-value</module>
   </modules>
 
   <dependencies>
