+++ pom.xml
@@ -64,9 +64,9 @@
   <modules>
     <module>resource</module>
     <module>manager</module>
+    <module>primitives</module>
     <module>collections</module>
     <module>coordination</module>
-    <module>atomic</module>
     <module>examples</module>
     <module>all</module>
   </modules>
