+++ primitives/src/main/resources/META-INF/services/io.atomix.catalyst.serializer.CatalystSerializable
@@ -1,0 +1,21 @@
+#
+# Copyright 2015 the original author or authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+io.atomix.primitives.state.ValueCommands$CompareAndSet
+io.atomix.primitives.state.ValueCommands$Get
+io.atomix.primitives.state.ValueCommands$GetAndSet
+io.atomix.primitives.state.ValueCommands$Set
+io.atomix.primitives.state.ValueCommands$Listen
+io.atomix.primitives.state.ValueCommands$Unlisten
